# docs/swagger.yaml
basePath: /
definitions:
...
  models.Trade:
    properties:
      closedAt:
        example: 1640999800
        type: integer
      createdAt:
        example: 1640995200
        type: integer
      entryPrice:
        example: 50000
        type: number
      error:
        example: ""
        type: string
      executedAt:
        example: 1640995260
        type: integer
      executedPrice:
        example: 50100.5
        type: number
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      leverage:
        example: 10
        type: integer
      orderId:
        example: 123456789
        type: integer
      pnl:
        example: 250.75
        type: number
      side:
        example: BUY
        type: string
      size:
        example: 1000
        type: number
      status:
        description: PENDING, ACTIVE, FILLED, CANCELED, FAILED
        example: ACTIVE
        type: string
      stopLoss:
        example: 49000
        type: number
      symbol:
        example: BTCUSDT
        type: string
      takeProfit:
        example: 52000
        type: number
      userId:
        example: user123
        type: string
    type: object
  models.TradeRequest:
    properties:
      entryPrice:
        example: 50000
        type: number
      leverage:
        example: 10
        maximum: 125
        minimum: 1
        type: integer
      side:
        description: '"BUY" or "SELL"'
        example: BUY
        type: string
      size:
        description: Position size in USDT
        example: 1000
        type: number
      stopLoss:
        example: 49000
        type: number
      symbol:
        description: e.g., "BTCUSDT"
        example: BTCUSDT
        type: string
      takeProfit:
        example: 52000
        type: number
      userId:
        example: user123
        type: string
    required:
    - entryPrice
    - leverage
    - side
    - size
    - stopLoss
    - symbol
    - takeProfit
    - userId
    type: object
  models.TradeResponse:
    properties:
      data: {}
      error:
        example: ""
        type: string
      message:
        example: Trade executed successfully
        type: string
      success:
        example: true
        type: boolean
      timestamp:
        example: 1640995200
        type: integer
      tradeId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@cryptotradingapi.com
    name: API Support
  description: |-
    Professional cryptocurrency trading API with automated Stop Loss and Take Profit
    Supports Binance Futures trading with real-time monitoring and Firebase logging
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Crypto Trading API
  version: "1.0"
paths:
...
  /api/trade:
    post:
      consumes:
      - application/json
      description: Execute a futures trade on Binance with stop loss and take profit
      parameters:
      - description: Trade parameters
        in: body
        name: trade
        required: true
        schema:
          $ref: '#/definitions/models.TradeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Trade executed successfully
          schema:
            $ref: '#/definitions/models.TradeResponse'
        "400":
          description: Invalid request or trade parameters
          schema:
            $ref: '#/definitions/models.TradeResponse'
        "401":
          description: Unauthorized - Invalid API key
          schema:
            $ref: '#/definitions/models.TradeResponse'
        "500":
          description: Internal server error - Trade execution failed
          schema:
            $ref: '#/definitions/models.TradeResponse'
      security:
      - ApiKeyAuth: []
      summary: Execute a new trade
      tags:
      - Trading
  /api/trade/{tradeId}:
    get:
      description: Retrieve a specific trade by its unique ID
      parameters:
      - description: Trade ID
        in: path
        name: tradeId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Trade retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/models.TradeResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Trade'
              type: object
        "401":
          description: Unauthorized - Invalid API key
          schema:
            $ref: '#/definitions/models.TradeResponse'
        "404":
          description: Trade not found
          schema:
            $ref: '#/definitions/models.TradeResponse'
      security:
      - ApiKeyAuth: []
      summary: Get trade by ID
      tags:
      - Trading
  /api/trades/{userId}:
    get:
      description: Retrieve all trades for a specific user ID
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Trades retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/models.TradeResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Trade'
                  type: array
              type: object
        "401":
          description: Unauthorized - Invalid API key
          schema:
            $ref: '#/definitions/models.TradeResponse'
        "500":
          description: Internal server error - Failed to fetch trades
          schema:
            $ref: '#/definitions/models.TradeResponse'
      security:
      - ApiKeyAuth: []
      summary: Get user trades
      tags:
      - Trading
...
securityDefinitions:
  ApiKeyAuth:
    description: Enter your API key to access protected endpoints
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"
...